Added support for player height & weight
TODO: when getting random heights and weights, be smart about it -- for
example, we don't want a whole bunch of 7'2" 150 lbs and 5'5" 350 lbs
guys. These should be much more rare than the general case. Like a bell
curve, we need weighted random numbers here.

+++ FootballSim/FootballSim/Models/Positions/EmptyPosition.cs
@@ -15,9 +15,9 @@
         {
             get { return Side.None; }
         }
-        public double MinWeight { get; set; }
-        public double MaxWeight { get; set; }
-        public double MinHeight { get; set; }
-        public double MaxHeight { get; set; }
+        public int MinWeight { get; set; }
+        public int MaxWeight { get; set; }
+        public int MinHeight { get; set; }
+        public int MaxHeight { get; set; }
     }
 }
